fix go lint error
