--- tests/test_runner.go
@@ -1,177 +1,0 @@
-package test
-
-import (
-	"encoding/json"
-	"flag"
-	"fmt"
-	"os"
-	"strings"
-	"testing"
-	"time"
-)
-
-// TestWithReporting runs tests and generates reports
-func TestWithReporting(m *testing.M) {
-	var generateReport bool
-	var reportFile string
-	var htmlFile string
-	
-	flag.BoolVar(&generateReport, "report", false, "Generate test report")
-	flag.StringVar(&reportFile, "report-file", "test-report.json", "Test report JSON file")
-	flag.StringVar(&htmlFile, "html-file", "test-report.html", "Test report HTML file")
-	flag.Parse()
-
-	if generateReport {
-		// Run tests with custom output
-		runTestsWithReport(m, reportFile, htmlFile)
-	} else {
-		// Run tests normally
-		os.Exit(m.Run())
-	}
-}
-
-func runTestsWithReport(m *testing.M, reportFile, htmlFile string) {
-	startTime := time.Now()
-	
-	// Capture test results
-	results := []TestResult{}
-	
-	// Run tests and capture output
-	exitCode := m.Run()
-	
-	endTime := time.Now()
-	
-	// Parse test output to extract results
-	// In a real implementation, you would parse the test output
-	// For now, we'll create sample results based on exit code
-	if exitCode == 0 {
-		results = append(results, TestResult{
-			Name:     "TestSESEmailVerification",
-			Status:   "PASS",
-			Duration: "5s",
-		})
-		results = append(results, TestResult{
-			Name:     "TestSESDomainVerification",
-			Status:   "PASS",
-			Duration: "10s",
-		})
-		results = append(results, TestResult{
-			Name:     "TestDMARCRecordValidation",
-			Status:   "PASS",
-			Duration: "3s",
-		})
-	} else {
-		results = append(results, TestResult{
-			Name:     "TestSESModule",
-			Status:   "FAIL",
-			Duration: "1s",
-			Error:    "Test failed",
-		})
-	}
-	
-	// Generate report
-	report := GenerateTestReport(results, startTime, endTime)
-	
-	// Save reports
-	if err := report.SaveReportToFile(reportFile); err != nil {
-		fmt.Fprintf(os.Stderr, "Failed to save JSON report: %v\n", err)
-	}
-	
-	if err := report.SaveReportToHTML(htmlFile); err != nil {
-		fmt.Fprintf(os.Stderr, "Failed to save HTML report: %v\n", err)
-	}
-	
-	// Print report to console
-	report.PrintReport()
-	
-	// Also write a simplified version for GitHub Actions
-	writeGitHubSummary(report)
-	
-	os.Exit(exitCode)
-}
-
-// writeGitHubSummary writes a GitHub-friendly summary for PR comments
-func writeGitHubSummary(report *TestReport) {
-	summaryFile := os.Getenv("GITHUB_STEP_SUMMARY")
-	if summaryFile == "" {
-		summaryFile = "test-summary.md"
-	}
-	
-	passRate := float64(report.PassedTests) / float64(report.TotalTests) * 100
-	
-	// Create markdown summary
-	var summary strings.Builder
-	summary.WriteString("## 🧪 Terraform AWS SES Test Results\n\n")
-	
-	// Status badge
-	if report.FailedTests == 0 {
-		summary.WriteString("![Tests Passed](https://img.shields.io/badge/tests-passed-success)\n\n")
-	} else {
-		summary.WriteString("![Tests Failed](https://img.shields.io/badge/tests-failed-critical)\n\n")
-	}
-	
-	// Summary table
-	summary.WriteString("### 📊 Summary\n\n")
-	summary.WriteString("| Metric | Value |\n")
-	summary.WriteString("|--------|-------|\n")
-	summary.WriteString(fmt.Sprintf("| Total Tests | %d |\n", report.TotalTests))
-	summary.WriteString(fmt.Sprintf("| ✅ Passed | %d |\n", report.PassedTests))
-	summary.WriteString(fmt.Sprintf("| ❌ Failed | %d |\n", report.FailedTests))
-	summary.WriteString(fmt.Sprintf("| ⏭️ Skipped | %d |\n", report.SkippedTests))
-	summary.WriteString(fmt.Sprintf("| 📊 Pass Rate | %.1f%% |\n", passRate))
-	summary.WriteString(fmt.Sprintf("| ⏱️ Duration | %s |\n\n", report.Duration))
-	
-	// Test details
-	summary.WriteString("### 📋 Test Details\n\n")
-	summary.WriteString("| Test Name | Status | Duration |\n")
-	summary.WriteString("|-----------|--------|----------|\n")
-	
-	for _, result := range report.Results {
-		status := ""
-		switch result.Status {
-		case "PASS":
-			status = "✅ Pass"
-		case "FAIL":
-			status = "❌ Fail"
-		case "SKIP":
-			status = "⏭️ Skip"
-		}
-		summary.WriteString(fmt.Sprintf("| %s | %s | %s |\n", result.Name, status, result.Duration))
-	}
-	
-	// Failed test details
-	if report.FailedTests > 0 {
-		summary.WriteString("\n### ❌ Failed Tests\n\n")
-		for _, result := range report.Results {
-			if result.Status == "FAIL" && result.Error != "" {
-				summary.WriteString(fmt.Sprintf("**%s**\n", result.Name))
-				summary.WriteString("```\n")
-				summary.WriteString(result.Error)
-				summary.WriteString("\n```\n\n")
-			}
-		}
-	}
-	
-	// Summary message
-	summary.WriteString("\n### 📝 Summary\n\n")
-	summary.WriteString(report.Summary + "\n")
-	
-	// Write to file
-	if err := os.WriteFile(summaryFile, []byte(summary.String()), 0644); err != nil {
-		fmt.Fprintf(os.Stderr, "Failed to write GitHub summary: %v\n", err)
-	}
-	
-	// Also write JSON for parsing in CI
-	reportData := map[string]interface{}{
-		"total":    report.TotalTests,
-		"passed":   report.PassedTests,
-		"failed":   report.FailedTests,
-		"skipped":  report.SkippedTests,
-		"passRate": passRate,
-		"summary":  report.Summary,
-	}
-	
-	if jsonData, err := json.Marshal(reportData); err == nil {
-		os.WriteFile("test-results.json", jsonData, 0644)
-	}
-}
